+++ modules/ec2/main.tf
@@ -15,14 +15,20 @@ provider "aws" {
  region                  = var.region
 }
 
-// the properties
+// the resources properties
 
 resource "aws_instance" "ec2" {
   instance_type          = var.instance_type
   ami                    = var.ami_type
   subnet_id              = var.subnet_id
   security_groups        = [var.security_groups_id]
+  user_data              = "${file("server-script.sh")}"
   tags = {
     Name                 = var.machine_name
   }
 }
+
+resource "aws_eip" "elastic_ip_addr" {
+  instance               = aws_instance.ec2.id
+}
+