+++ CHANGELOG.md
@@ -1,8 +1,18 @@
-# ChangeLog (status = added + fixed)
+# ChangeLog (status = added)
 
 *Q.how to read changelog file ?* 
 
 *A.the correct reading order is from the bottom to top*
+#
+
+- **14/07/2023** 
+	> checking ec2 module = **code works!**
+
+	> adding "user_data" (web-server script) to ec2 module
+
+	> adding output file to ec2 module
+
+	> adding "EIP" resource to ec2 module
 
 - **07/07/2023** 
 	> checking ec2 & s3-bucket modules = **code works!**